Comments and restructure
